requirements.txt
numpy
torch
opencv-python
argparse
pillow
tqdm
scikit-learn
matplotlib
